fix: Implemented Generalization of Switches to ToggleableView
Resolves: #5
